refactor(app): extract headless-shared from headless.ts
